jbd2: Increase the ext4 default commit age
Both the VM and EXT4 have a "commit to disk after X seconds" time.
Currently the EXT4 time is shorter than our VM time, which is a bit
suboptional,
it's better for performance to let the VM do the writeouts in bulk
rather than something deep in the journalling layer.

Change-Id: I0de7bcfa9487421954cb9ae7f2cab3159c663878
Signed-off-by: Arjan van de Ven <arjan@linux.intel.com>
Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>
Signed-off-by: onettboots <blackcocopet@gmail.com>