serial: msm_geni_serial: memset RX buffer to Zero
Currently driver allocating RX buffer once in port open and use the
same buffer for each RX transfer. This buffer is not cleared in driver.
In some scenarios when SW got DMA_DONE interrupt and the RX buffer we are
still seeing the previously completed RX data in DDR buffer. DMA_DONE from
HW doesn't confirm that the DATA is copied to DDR, sometimes we are queuing
the stale data from previous transfer to tty flip_buffer.

This change is to memset RX buffer to zero after tty_flip_buffer_push
so that memset change will easily help us to identify suck scenarios.

Change-Id: If14bdf0377cd6764280f515b1b8e67f5629e32ce
Signed-off-by: Chandana Kishori Chiluveru <cchiluve@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>