kernel/sys.c: Strip localversion from uname for GMS processes
Play Integrity now hashes the kernel's localversion and compares it to a
blacklist of localversion hashes for the purpose of failing Play Integrity
attestation.

Among the banned localversion hashes are "-Sultan" and "-sultan".

Since this is biased and doesn't improve security for end users in any way,
strip the localversion from uname for processes which are part of GMS's
thread group so that Play Integrity doesn't fail due to this.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: kyvangka1610 <kyvangka2002@gmail.com>
Signed-off-by: onettboots <blackcocopet@gmail.com>